feat: collection service to map setRule
